Refactoring of the disassembler.
