+++ smwe_rom/src/disassembler/instruction.rs
@@ -1,4 +1,4 @@
-use std::fmt::Display;
+use std::fmt::{Display, Formatter};
 
 use smallvec::{smallvec, SmallVec};
 
@@ -15,23 +15,25 @@ use crate::{
 
 #[derive(Copy, Clone, Eq, PartialEq, Debug)]
 pub struct Instruction {
+    pub offset: AddrPc,
     pub opcode: Opcode,
+    pub m_flag: bool,
+    pub x_flag: bool,
+
     // Length might be shorter than 4, needs to be looked up by opcode
-    operands:   [u8; 4],
+    operands: [u8; 4],
 }
 
-#[derive(Copy, Clone)]
-pub struct DisplayInstruction {
-    i:      Instruction,
-    offset: AddrPc,
-    x_flag: bool,
-    m_flag: bool,
-}
+#[derive(Copy, Clone, Eq, PartialEq, Debug)]
+pub struct DisplayInstruction(Instruction);
+
+#[derive(Copy, Clone, Eq, PartialEq, Debug)]
+pub struct DisplayInstructionWithFlags(Instruction);
 
 // -------------------------------------------------------------------------------------------------
 
 impl Instruction {
-    pub fn parse(bytes: &[u8], p_reg: PRegister) -> Result<(Self, &[u8]), InstructionParseError> {
+    pub fn parse(bytes: &[u8], offset: AddrPc, p_reg: PRegister) -> Result<(Self, &[u8]), InstructionParseError> {
         let (&opcode_raw, rest) = bytes.split_first().ok_or(InstructionParseError::InputEmpty)?;
         let mut opcode = SNES_OPCODES[opcode_raw as usize];
 
@@ -49,11 +51,15 @@ impl Instruction {
         let mut operands = [0u8; 4];
         operands[..operands_v.len()].copy_from_slice(operands_v);
 
-        Ok((Self { opcode, operands }, rest))
+        Ok((Self { offset, opcode, operands, m_flag: p_reg.m_flag(), x_flag: p_reg.x_flag() }, rest))
     }
 
-    pub fn display(self, offset: AddrPc, x_flag: bool, m_flag: bool) -> DisplayInstruction {
-        DisplayInstruction { i: self, offset, x_flag, m_flag }
+    pub fn display(self) -> DisplayInstruction {
+        DisplayInstruction(self)
+    }
+
+    pub fn display_with_flags(self) -> DisplayInstructionWithFlags {
+        DisplayInstructionWithFlags(self)
     }
 
     pub fn operands(&self) -> &[u8] {
@@ -94,7 +100,7 @@ impl Instruction {
                 smallvec![]
             }
             _ => {
-                if self.opcode.mnemonic.can_branch() {
+                if self.can_branch() {
                     log::error!("Unhandled branching instruction {self:?} at ${offset:06X}");
                     smallvec![]
                 } else {
@@ -143,21 +149,40 @@ impl Instruction {
             _ => 0,
         } as usize)
     }
+
+    pub fn can_branch(self) -> bool {
+        self.opcode.mnemonic.can_branch()
+    }
+
+    pub fn is_subroutine_call(self) -> bool {
+        self.opcode.mnemonic.is_subroutine_call()
+    }
+
+    pub fn is_subroutine_return(self) -> bool {
+        self.opcode.mnemonic.is_subroutine_return()
+    }
+}
+
+impl Display for DisplayInstructionWithFlags {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        write!(f, "[{}{}] ", ['m', 'M'][self.0.m_flag as usize], ['x', 'X'][self.0.x_flag as usize])?;
+        self.0.display().fmt(f)
+    }
 }
 
 impl Display for DisplayInstruction {
-    fn fmt(&self, outer_fmt: &mut std::fmt::Formatter) -> std::fmt::Result {
+    fn fmt(&self, outer_fmt: &mut Formatter) -> std::fmt::Result {
         use std::io::Write;
         let mut fmt: SmallVec<[u8; 64]> = Default::default();
-        let DisplayInstruction { i, x_flag, m_flag, offset } = *self;
+        let Instruction { x_flag, m_flag, offset, .. } = self.0;
 
         let offset = AddrSnes::try_from_lorom(offset).unwrap_or_default();
-        let address_long = i.get_intermediate_address(offset).0;
+        let address_long = self.0.get_intermediate_address(offset).0;
         let address_short = address_long & 0xFFFF;
         let address_dp = address_long & 0xFF;
 
-        write!(fmt, "{}", i.opcode.mnemonic).unwrap();
-        match i.opcode.mode {
+        write!(fmt, "{}", self.0.opcode.mnemonic).unwrap();
+        match self.0.opcode.mode {
             Implied => {
                 // no-op
             }
@@ -165,30 +190,30 @@ impl Display for DisplayInstruction {
                 write!(fmt, " A").unwrap();
             }
             Constant8 | Immediate8 => {
-                write!(fmt, " #${:02X}", i.operands[0]).unwrap();
+                write!(fmt, " #${:02X}", self.0.operands[0]).unwrap();
             }
             Immediate16 => {
-                write!(fmt, " #${:04X}", u16::from_le_bytes([i.operands[0], i.operands[1]])).unwrap();
+                write!(fmt, " #${:04X}", u16::from_le_bytes([self.0.operands[0], self.0.operands[1]])).unwrap();
             }
             ImmediateXFlagDependent | ImmediateMFlagDependent => {
-                let x = i.opcode.mode == ImmediateXFlagDependent && x_flag;
-                let m = i.opcode.mode == ImmediateMFlagDependent && m_flag;
+                let x = self.0.opcode.mode == ImmediateXFlagDependent && x_flag;
+                let m = self.0.opcode.mode == ImmediateMFlagDependent && m_flag;
                 if x || m {
-                    write!(fmt, " #${:02X}", i.operands[0]).unwrap();
+                    write!(fmt, " #${:02X}", self.0.operands[0]).unwrap();
                 } else {
-                    write!(fmt, " #${:04X}", u16::from_le_bytes([i.operands[0], i.operands[1]])).unwrap();
+                    write!(fmt, " #${:04X}", u16::from_le_bytes([self.0.operands[0], self.0.operands[1]])).unwrap();
                 }
             }
             DirectPage => {
                 write!(fmt, " ${address_dp:02X}").unwrap();
             }
             Relative8 => {
-                let address = i.operands[0] as u32;
+                let address = self.0.operands[0] as u32;
                 let address = address & !(-1i32 << 8) as u32;
                 write!(fmt, " ${address:02X}").unwrap();
             }
             Relative16 => {
-                let address = u16::from_le_bytes([i.operands[0], i.operands[1]]) as u32;
+                let address = u16::from_le_bytes([self.0.operands[0], self.0.operands[1]]) as u32;
                 let address = address & !(-1i32 << 16) as u32;
                 write!(fmt, " ${address:04X}").unwrap();
             }
@@ -235,7 +260,7 @@ impl Display for DisplayInstruction {
                 write!(fmt, " [${address_dp:02X}], Y").unwrap();
             }
             BlockMove => {
-                write!(fmt, " ${:02X}, ${:02X}", i.operands[0], i.operands[1]).unwrap();
+                write!(fmt, " ${:02X}, ${:02X}", self.0.operands[0], self.0.operands[1]).unwrap();
             }
         };
         outer_fmt.pad(std::str::from_utf8(&fmt).unwrap())
